Results module recoded from scratch.
There are two basic tables, frames and events, and their respective
mappers, Frame and Event. A Frame row represents the equivalent of
a SDSS frame fits file. Event represents a single line on a single
frame. Multiple Events can occur on a single Frame constituting a
many to one relationship between the two.

Because many of the columns in the Event are parameters to objects
(i.e. a Line and a Time object) they are declared to be composite
columns and are mapped by those specific objects that then
implement the actual logic and calculation functionality (f.e.
coordinate transformations, time format transformations etc.).

This is a big step away from the flat table used so far and will
hopefully make it easier to use and understand in the future. The
abstraction has changed drastically but it came with a price -
spaghetti code. SQLAlchemy can't figure out when an single key of
a composite foreign key has changed and does not cascade the changes
back through the tables so all the classes have to supply getters and
setters that maintain data integrity which is a lot of work to do
consistently. Additional penalty is that the interface does not look
consistent - there's several obvious ways you can do the same thing,
which is unpythonic.