Added support to change time stamps and have them persist to the DB.
The interface is again not consistent with how the Point is handled.
The problem is that line-time is actually two time stamps and it's
ambiguous as to which one (start or end time) needs to be changed.

If you let people walk all the way down to the BasicTime instance
in LineTime, like it's done for Point, then sqlalchemy can not re-
gister that the session became dirty and unfortunatelly there is
no easy way to deal with that.

So currently the time is chnaged over methods for which the user
has to instantiante a new BasicTime or LineTime object and then
consistency is handled within the setter methods.