LFDS renamed to lfd and declared a package. Changes in how DetectTrails core module interacts with the environment. Changes in the entry point and setup of the package.
Declaring everything a package helps a lot with imports and code clea-
nliness. The environment is now set up from the init script by an
explicit call to the setup function by the user in the case the
package wants to be used interactively.
The environment can be set up by the user in the terminal in which
case the code just expects the paths to exist.

Results and errors file access is now changed to a context manager.
This aleviates a lot of issues with interactive use of the package
but still leaves the files open during the whole batch processing
duration.

Changes to how debugging save paths - consequence of a different
way the package recognizes the environment.