Results module rewrite finished, awaiting documentation.
Ok, so again another results module rewrite because of incosistencies
and lacking documentation. Main changes consist of declaring
PRAGMA FOREIGN_KEYS=ON during DB engine declaration. There are several
advantages to this. Firstly, commiting to the DB is now made much
easier because the foreign key depenedencies are auto-commited to
the DB if linked to the parent table. Secondly, 'cascade' keywords
now work as expected and on parent update all children ae updated too.
One problem here was sqlite DB does not seem to want to happily support
referential integrity so children would not be updated on parent change
without setting 'passive_updates' (on the parent) to False. This fixes
that issues at the cost of decreased performance because each child is
queried for a change on parent update.

BasicTime has been upgraded to a TypeDecorator, basically shortcutting
the stored Float and instantiating an Astropy Time object from values
queried from the DB. On commit action the Time object is converted to
SDSS TAI format and stored. None values are permitted because of the
line start and end times but the column in the Frame table has been
declared not nullable.

Point object has been upgraded to a MutableComposite and now has
hooks to all on-change events. The events triggering the change of the
coordinates in the frame coord sys. will trigger an update to parent
table. So technically it is possible to change the ccd coord sys
coordinates directly by accessing the __c* values and end up in an
inconsistent situation. This should never happen however because all
getter and setter actions are overridden and they always re-instantiate
all of the values to make sure consistency is preserved. Unfortunately,
it seems impossible to trigger an update of composite object on update
of one of its components directly in the table so the getters and setters
for the table itself have to keep that consistent as well, thus no reduction
of code/line-count is achieved, but I believe this is more in the SQLAlch.
spirit so there we go.

TODO: Write all documentation for the results module and write a lot of
examples. Get a docs site working because honestly it's just way to easy
to forget what I thought of or imagined how it should work as. Write
some unit tests for this whole thing please! Get the GUI working. See if
Create-Jobs module is working, update the docs there. Check if docs are
up to date for the detecttrails module. Get the GUI working with the new
results module (a session context manager?) and check the docs there
(but they should be ok for the most part).