This is the first version of results module that works and is well documented.
The large Line class was taken out as it just adds a layer of unnecessary
complexity. Instead the old approach with the Point class was re-adopted.
The Point class however was extended to constinually follow all the values
of coordinates and has allowed user to access all the values of coordinates
all the time. Inconsistencies are only warned about.

The Event class was extended and wraps all the attributes in properties now.
This is verbose and likely unnecessary but is currently the only way to get
things working like I want them.

There have been so many attempts at various things that it's hard to say what
is everything that has been done. SQL and DBs are hard.

TODO: add a pretty print code to utils. Document utils. Sleep. Try not to die.