Provisional conversion of the imagechecker module to python 3.
Major refitting of the Images class to get it to work with the
SQLAlchemy's ORM functionality from the results module.
Images class got renamed to Data class to match reality better.
The results module was altered by adding to additional columns to
events table (verified and false positive) in order to keep track
of whether individual Event was ever manually inspected (passed or
failed).
The rework of the module has all the basic features working but is
not complete. A complete redesign of the Data class is required so
that the results data (Events) can be separated from the Images,
because that's the only way to handle the case when someone just
wants to browse through the images themselves. Additional opportunity
then is opened to create such an Images class to allow different dirs
of images in different locations to be included etc. This requires
additional code.
Added documentation for the classes in imagechecker module.
Cleaned up old functionality in utils module and added a new any-depth
getattr function.