Complete recoding of the imagechecker module. The new working idea is that the imagechecker gui itself does not handle any kind of indexing, browsing or fact-checking logic beacuse that was too messy. Now the databases are indexed via Indexers. Each indexers orders the rows from the database and enables user to browse through them iteratively, or by skipping, select a particular item from them or jump to a particular index or unique id of the ordered items.
The consistency between Indexers are kept track of in a Browser. A
Browser makes sure that when the primary Indexer moves, jumps or skips
the secondary indexer moves, jumps or skips to the correct object and
displays the correct, matching, data. There are two defined Browsers
for the dataset in hand - EventBrowser and ImageBrowser. Event browser
can be used to browse all events (even if the ImageIndexer is a sub or
superset of all Events indexed by EventIndexer and vice-versa).

The results module recieved a minor change - the cascade has been set
to trigger on 'expunge' actions as well, but it does not seem to work.