Rewritten, refactored and (very!) optimized profiles module.
Declared __all__ for modules with a lot of functionality.
Standardized the output of the 'f' functions of Sources, Seeing...

Refactored the convolutions functions - there are now two convolution
functions that work on specific number of objects (two and three). The
other functions, f.e. convolve_defocus/seeing..., are now aliases to
those specific convolutions. The general recursive convolution function
is now checking the number of objects to convolve and delegating to
the specific convolutions instead of blindly convolving recoursively.

All Sources have standardized output type of their self.f() functions.
These functions have also been recoded significantly to increase the,
what was very poor, performance. Optimizations allowed from 2x to 1400x
speedups, mainly through vectorization of the problem and removal of map
and loops. Cumulatively this can be felt significantly when convolving
a lot of objects with large scales and especially when using rescale().

Added simple plot utility. this still has to be filled in with practical
functions, but at least as a placeholder for now.

Added a series of premade rabina profiles. Profiles span the range of
0-1.5 radians (0-86 degrees) in increments of 0.1 radians (5 degrees).
The rotation corresponds to the angle between the meteors direction of
travel and line of sight. Code required to produce these profiles is
provided but unmaintained. The rotation matrix creation was very slow and
implemented in weave which is not officially supported anymore. There is
a Python version of this function supplied, but performance unknown.

Added bunch of docstrings and comments.
Moved the commonly used constants to a const.py module.