Changed my mind. Indexer class should not depend on the dbconf dictionary.
Instead subclasses can override the crucial methods and code that
up. This should increase readability, maintainability but most
importantly it increases flexibiity. Specific Indexers can now be
defined over a subset of all rows in the database by selecting only
rows of interest and assigning them to items in the initFromDB method.